+++ packages/terra-compact-interactive-list/src/index.js
@@ -1,0 +1,5 @@
+import CompactInteractiveList from './CompactInteractiveList';
+import { widthUnitTypes, alignTypes } from './utils/constants';
+
+export default CompactInteractiveList;
+export { widthUnitTypes, alignTypes };