+++ packages/terra-table/src/Table.jsx
@@ -24,6 +24,11 @@ import getFocusableElements from './utils/focusManagement';
 
 const cx = classNames.bind(styles);
 
+export const rowSelectionModes = {
+  SINGLE: 'single',
+  MULTIPLE: 'multiple',
+};
+
 const propTypes = {
   /**
    * Unique id used to identify the table.
@@ -138,10 +143,10 @@ const propTypes = {
   onRowSelectionHeaderSelect: PropTypes.func,
 
   /**
-   * Boolean indicating whether or not the table should allow entire rows to be selectable. An additional column will be
-   * rendered to allow for row selection to occur.
+   * Enables row selection capabilities for the table.
+   * Use 'single' for single row selection and 'multiple' for multi-row selection.
    */
-  hasSelectableRows: PropTypes.bool,
+  rowSelectionMode: PropTypes.oneOf(Object.values(rowSelectionModes)),
 
   /**
    * Boolean indicating whether or not the table columns should be displayed. Setting the value to false will hide the columns,
@@ -191,12 +196,12 @@ function Table(props) {
     defaultColumnWidth,
     columnHeaderHeight,
     rowHeight,
+    rowSelectionMode,
     onColumnSelect,
     onCellSelect,
     onSectionSelect,
     onRowSelect,
     onRowSelectionHeaderSelect,
-    hasSelectableRows,
     hasVisibleColumnHeaders,
     isStriped,
     rowHeaderIndex,
@@ -254,6 +259,7 @@ function Table(props) {
     isResizable: false,
   };
 
+  const hasSelectableRows = rowSelectionMode === rowSelectionModes.MULTIPLE;
   const displayedColumns = (hasSelectableRows ? [tableRowSelectionColumn] : []).concat(pinnedColumns).concat(overflowColumns);
   const [tableColumns, setTableColumns] = useState(displayedColumns.map((column) => initializeColumn(column)));
 
@@ -306,11 +312,11 @@ function Table(props) {
     }
 
     // Since the row selection mode has changed, the row selection mode needs to be updated.
-    setRowSelectionModeAriaLiveMessage(intl.formatMessage({ id: hasSelectableRows ? 'Terra.table.row-selection-mode-enabled' : 'Terra.table.row-selection-mode-disabled' }));
+    setRowSelectionModeAriaLiveMessage(intl.formatMessage({ id: rowSelectionMode === rowSelectionModes.MULTIPLE ? 'Terra.table.row-selection-mode-enabled' : 'Terra.table.row-selection-mode-disabled' }));
 
     setTableColumns(displayedColumns.map((column) => initializeColumn(column)));
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [hasSelectableRows]);
+  }, [rowSelectionMode]);
 
   // useEffect for row updates
   useEffect(() => {
@@ -581,10 +587,10 @@ function Table(props) {
               text={section.text}
               rows={section.rows}
               rowHeight={rowHeight}
-              hasRowSelection={hasSelectableRows}
+              rowSelectionMode={rowSelectionMode}
               displayedColumns={displayedColumns}
               rowHeaderIndex={rowHeaderIndex}
-              onCellSelect={isGridContext || hasSelectableRows ? handleCellSelection : undefined}
+              onCellSelect={isGridContext || rowSelectionMode ? handleCellSelection : undefined}
               onSectionSelect={onSectionSelect}
             />
           ))}