+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added
+  * Added examples and tests to `terra-table` for single row selection mode.
   * Added examples for resizable columns in tables.
 
 ## 1.45.0 - (November 20, 2023)
@@ -12,6 +13,7 @@
   * Added new test for Terra Slider component for long field labels.
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
   * Added examples and tests to `terra-data-grid` for sections in FlowsheetDataGrid.
+  * Added documentation and examples for `terra-compact-interactive-list`.
   
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.